+++ .gitignore
@@ -1,5 +1,6 @@
 .gradle
-/build/
+**/build/
+**/out/
 
 *.jar
 !gradle-wrapper.jar
