+++ core/build.gradle
@@ -1,0 +1,17 @@
+plugins {
+  id 'nebula.kotlin' version '1.1.4-2'
+}
+
+apply {
+  from "$rootDir/gradle/kotlin.gradle"
+}
+
+repositories {
+  mavenCentral()
+  jcenter()
+}
+
+dependencies {
+  compile "com.fasterxml.jackson.module:jackson-module-kotlin:${Dependencies.JACKSON}"
+  compile "com.fasterxml.jackson.datatype:jackson-datatype-jdk8:${Dependencies.JACKSON}"
+}
