+++ core/src/main/kotlin/com/github/ptrteixeira/cookbook/core/User.kt
@@ -1,0 +1,7 @@
+package com.github.ptrteixeira.cookbook.core
+
+import java.security.Principal
+
+data class User(val id: String): Principal {
+    override fun getName(): String  = id
+}
