+++ server/src/main/kotlin/com/github/ptrteixeira/cookbook/CookbookConfiguration.kt
