+++ server/src/main/kotlin/com/github/ptrteixeira/cookbook/auth/AuthComponent.kt
@@ -1,0 +1,17 @@
+package com.github.ptrteixeira.cookbook.auth
+
+import com.github.ptrteixeira.cookbook.base.BaseComponent
+import com.github.ptrteixeira.cookbook.core.User
+import dagger.Component
+import io.dropwizard.auth.Authenticator
+import javax.inject.Named
+
+@Component(modules = arrayOf(AuthModule::class), dependencies = arrayOf(BaseComponent::class))
+interface AuthComponent {
+    @Named(TOKEN_AUTH)
+    fun tokenAuth(): Authenticator<String, User>
+
+    companion object {
+        const val TOKEN_AUTH = "com.github.ptrteixeira.cookbook.auth.tokenAuth"
+    }
+}
