+++ server/src/main/kotlin/com/github/ptrteixeira/cookbook/base/BaseComponent.kt
