+++ server/src/main/kotlin/com/github/ptrteixeira/cookbook/config/OauthConfiguration.kt
