+++ server/src/main/kotlin/com/github/ptrteixeira/cookbook/data/DataComponent.kt
