+++ server/src/main/kotlin/com/github/ptrteixeira/cookbook/resources/ResourcesComponent.kt
