+++ server/src/test/resources/fixtures/recipe-egg-absent-field.json
