+++ server/src/test/resources/fixtures/recipe-egg.json
