--- src/main/kotlin/com/github/ptrteixeira/cookbook/model/User.kt
@@ -1,7 +1,0 @@
-package com.github.ptrteixeira.cookbook.model
-
-import java.security.Principal
-
-data class User(val id: String): Principal {
-    override fun getName(): String  = id
-}
